local_o1_agents/requirements_minimal.txt
# Essential dependencies for Local O1 integration
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
ollama>=0.1.0
streamlit>=1.28.0
psutil>=5.9.0
requests>=2.28.0
numpy>=1.24.0
pillow>=9.0.0
datasets>=2.14.0
matplotlib>=3.7.0
networkx>=3.1.0
pyvis>=0.3.2
memory-profiler>=0.61.0
torchaudio>=2.0.0
